split basic functions
