+++ oleveler/RNASeq_processor/RNASeq_loader.py
@@ -3,7 +3,6 @@ import os
 import pandas as pd
 import numpy as np
 
-# from oleveler.project_logger import logger
 from oleveler import calHash, logger
 from .biology_information import _parseGff, _parseTableForGeneLength
