+++ oleveler/__init__.py
@@ -1,3 +1,5 @@
+import rpy2.robjects as robjects # run this here or in .main can supress INFO: "R is already initialized. No need to initialize."
+
 from .main import *
 
 from .side_notebook_functions import \
@@ -9,3 +11,5 @@ from time import sleep
 
 from .MaxQuant_data_processor import *
 from .RNASeq_processor import *
+from .metadata_processor import loadMeta, removeExperiments, removeIds
+from .basic_stats import getStats
