+++ oleveler/main.py
@@ -31,239 +31,11 @@ from collections import OrderedDict
 from hashlib import md5
 from itertools import cycle
 import json
-import pickle
 import os
 import itertools
 
 from .project_logger import logger
 from .side_notebook_functions import current_jupytertheme_style
-from .safe_annotations import safeCol, safeAnnotations, safeMQdata
-
-# ==================== LOAD DATA =======================
-def loadMeta(path, toRemove=[]):
-    '''
-    Usage: 
-        metaDf, conditions, experiments = loadMeta('Annotation.csv')
-
-    Read metadata file. Now using the criteria of MSstats:
-    Columns of the metadata file (csv):
-    Raw.file,Condition,BioReplicate,Experiment
-    Thes columns can be used in MSstats analysis
-    This will be further translated to a DESeq2 required file on the fly.
-
-    The difference between metadata files used by MSstats and DESeq2 is that
-    DESeq2 accepts multiple factor table while MSstats only accepts one factor: Condition.
-    This makes meta data transformation easier from MSstats to DESeq2. 
-
-
-    Return metaDf, conditions, experiments
-
-    metaDf => pd.DataFrame
-        indexs = names of raw files, DO NOT include extension
-        cols = Raw.file, Condition, BioReplicate, Experiment
-            eg.
-                Condition = strain+condition
-                BioReplicate = 1,1,1,2,2,2,3,3,3
-                Experiment = strain+condition+repnumber (single experiment)
-    conditions => list
-    experiments => dict(cond1:[exp1, exp2, exp3], cond2:[exp4, exp5, exp6]...)
-    '''
-    logger.info(f'####### Load metadata #######')
-    logger.info(f'Metadata path: {path}')
-    metaDf = pd.read_csv(path, index_col=0)
-    assert not metaDf.index[0].endswith('.raw'), f'Please remove ".raw" from table {path}'
-    metaDf.loc[:,"Experiment"] = safeCol(metaDf['Experiment'])
-    metaDf.loc[:,"Condition"] = safeCol(metaDf['Condition'])
-    toRemove = safeCol(toRemove)
-    metaDf = metaDf[~metaDf['Experiment'].isin(toRemove)]
-    conditions = list(set(metaDf.Condition.to_list()))
-    conditions.sort()
-    logger.info('All conditions: ' + ', '.join(conditions))
-
-    experiments = OrderedDict()
-    for c in conditions:
-        exps = metaDf.Experiment[metaDf.Condition == c].to_list()
-        exps.sort()
-        experiments[c] = exps
-    logger.info(f'####### END Load metadata #######\n')
-
-    return metaDf, conditions, experiments
-
-
-# ====================== DATAP ROCESSING ======================
-
-# Basic stats
-
-def removeExperiments(metaDf, experiments, toRemove, dataDf=None):
-    # TODO update
-    # ??
-    """
-    Usage:
-        newMetaDf, newConditions, newExperiments, newDataDf = removeExperiments(metaDf, experiments, ['condA_rep_1', 'condB_rep_3'], dataDf)
-    toRemove should be a list of experiment names (columns of the data table) to remove
-    Provided list should have exact match.
-    experiments => dict(cond1:[exp1, exp2, exp3], cond2:[exp4, exp5, exp6]...)
-    """
-    logger.info(f'Remove {toRemove} from experiments')
-    newExps = OrderedDict()
-    newMetaDf = pd.DataFrame()
-    for c in experiments:
-        exps = [e for e in experiments[c] if e not in toRemove]
-        if len(exps) > 0:
-            newExps[c] = exps
-            newMetaDf = pd.concat((newMetaDf, metaDf.loc[metaDf['Experiment'].isin(exps), :]), axis=0)
-
-    if isinstance(dataDf, type(None)):
-        newDataDf = None
-    else:
-        newDataDf = dataDf.loc[:, dataDf.columns.isin(toRemove)]
-    newConditions = list(newExps.keys())
-
-    return newMetaDf, newConditions, newExps, newDataDf
-
-
-def removeIds(dataDf, toRemove):
-    """Usage:
-        newDf = removeIds(dataDf, ['idx1', 'idx2', ...])
-
-    Args:
-        dataDf (pd.DataFrame): input data table
-        toRemove (list): list of ids to remove. Will consider partial match.
-
-    Returns:
-        newDf: removed df
-    """
-    logger.info(f'Remove genes/proteins in this list {toRemove}')
-    newDf = dataDf.copy()
-    newIdx = [i for i in newDf.index if all(s not in i for s in toRemove)]
-    removed = newDf.index.difference(newIdx)
-    logger.info(f'Actual ids removed: {list(removed)}')
-    newDf = newDf.loc[newIdx, :]
-    return newDf
-
-
-def getStats(dataDf, experiments, title=''):
-    """
-    Usage: meanDf, nquantDf, varDf, stdDf, semDf = getStats(lfqDf, experiments)
-    If you want to exclude any data, do that before passing data in here.
-    """
-    # calculate hash for parameters
-    ha = calHash(dataDf, experiments)  # skipping title for now
-
-    os.makedirs('dataTables', exist_ok=True)
-
-    meanTbFile = f'dataTables/mean_{title}_{ha}.tsv'.replace('__', '_')
-    nquantTbFile = f'dataTables/nquant_{title}_{ha}.tsv'.replace('__', '_')
-    varTbFile = f'dataTables/var_{title}_{ha}.tsv'.replace('__', '_')
-    stdTbFile = f'dataTables/std_{title}_{ha}.tsv'.replace('__', '_')
-    semTbFile = f'dataTables/sem_{title}_{ha}.tsv'.replace('__', '_')
-
-    outputFiles = [meanTbFile, nquantTbFile, varTbFile, stdTbFile, semTbFile]
-
-    if all(os.path.isfile(f) for f in outputFiles):
-        outputDfs = [pd.read_csv(f, sep='\t', index_col=0) for f in outputFiles]
-        [logger.info(f'Read basic stats from {f}') for f in outputFiles]
-        return outputDfs
-    conditions = list(experiments.keys())
-    logger.info('Calculate stats with ddof at least 1 (ignoring the conditions with only one quantification).\n')
-    resDf = pd.DataFrame(index=dataDf.index, columns=conditions)
-    meanDf = resDf.copy()
-    stdDf = resDf.copy()
-    nquantDf = resDf.copy()
-    semDf = resDf.copy()
-    varDf = resDf.copy()
-
-    for c in conditions:
-        data = dataDf.loc[:, experiments[c]]
-
-        mean = np.nanmean(data, axis=1)
-        n = (~np.isnan(data)).sum(axis=1)
-        var = np.nanvar(data, axis=1, ddof=1)
-        std = np.sqrt(var)
-        sem = std / np.sqrt(n)
-        for df, x in zip(
-            [meanDf, nquantDf, varDf, stdDf, semDf],
-            [mean, n, var, std, sem]
-        ):
-            df.loc[:, c] = x
-
-    for df, f in zip([meanDf, nquantDf, varDf, stdDf, semDf], outputFiles):
-        df.to_csv(f, sep='\t')
-        logger.info(f'Write basic stats data to {f}.')
-
-    return meanDf, nquantDf, varDf, stdDf, semDf
-
-def genComparisonResults(compResultFile, comparisons):
-    """
-    allComparisons - dict of {'comp1':DF, 'comp2':DF,...}
-
-    DF.columns = ['log2FC', 'SE', 'pvalue', 'adj.pvalue', 'ImputationPercentage']
-    for MSstats result
-    DF.columns = ['baseMean', 'log2FC', 'SE', 'pvalue', 'adj.pvalue']
-    for DESeq2 result
-    """
-    compResultDf = pd.read_csv(compResultFile, sep='\t', index_col=0)
-
-    if 'ImputationPercentage' in compResultDf.columns:
-        # MSstats output
-        # "Label"	"log2FC"	"SE"	"Tvalue"	"DF"	"pvalue"	"adj.pvalue"	"issue"	"MissingPercentage"	"ImputationPercentage"
-        targetCols = ['log2FC', 'SE', 'pvalue', 'adj.pvalue', 'ImputationPercentage']
-    elif 'baseMean' in compResultDf.columns:
-        # DESeq2 result
-        # "Label" is added while combining all data
-        # "baseMean"	"log2FoldChange"	"lfcSE"	"pvalue"	"padj"
-        # for shrinked data, note the column "lfcSE" is actually "posterior SD"
-        # OR
-        # "baseMean"	"log2FoldChange"	"lfcSE"	"stat"	"pvalue"	"padj"
-        # for non-shrinked data
-        targetCols = ['log2FC', 'SE', 'pvalue', 'adj.pvalue', 'baseMean']
-    else:
-        raise ValueError("Do not know result type, make sure 'ImputationPercentage' or 'baseMean' in input data columns")
-
-    allCompResults = {}
-    for c in comparisons:
-        compData = compResultDf[compResultDf.Label == c]
-        compData = compData.loc[:, targetCols]
-        allCompResults[c] = compData
-
-    return allCompResults
-
-
-def _checkExistingCompResult(compExcel, sourceDf, compResultFileBase):
-    # compResultFile => path, see if the file exists
-    comparisons = pd.read_excel(compExcel, index_col=0).T.to_dict()
-    allCompResults = None
-    with open(compExcel, 'rb') as f:
-        cur_compExcelHash = md5(f.read()).hexdigest()
-    cur_sDfHash = md5(sourceDf.to_json().encode()).hexdigest()
-    compResultFile, ext = os.path.splitext(compResultFileBase)
-    compResultFile += f'_{cur_compExcelHash[:6]}_{cur_sDfHash[:6]}{ext}'
-    if os.path.isfile(compResultFile):
-        logger.info(f'Current {compExcel} hash = {cur_compExcelHash}.')
-        logger.info(f'Current input table hash = {cur_sDfHash}.')
-        logger.info(f'They should match previous result {compResultFile}.')
-        # check if current comparisons.xlsx match previous
-        allCompResults = genComparisonResults(compResultFile, comparisons)
-    return allCompResults, comparisons, compResultFile
-
-
-def getSig(compDf, tFc=1.5, tPv=0.05):
-    """return compDf with only significant ids
-
-    Note if you need significant ids in only one experiment, do not pass whole compDf
-
-    Args:
-        compDf ([type]): [description]
-        tFc (float, optional): [description]. Defaults to 1.5.
-        tPv (float, optional): [description]. Defaults to 0.05.
-
-    Returns:
-        compDf: filtered
-    """
-    f = compDf['log2FC'] > np.log2(tFc)
-    f = f | (compDf['log2FC'] < -np.log2(tFc))
-    f = f & (compDf['adj.pvalue'] < tPv)
-    return compDf[f]
 
 
 def grayout(c, scale=1.8):
